Preparing to test 3denvar with ensemble spun off into separate routine
Noticed that can make massive speadup in tuning with re-used ensembles
for background when not cycling perturbations, going to test if this can
be run and begin spinnin additional simulations.

+++ scripts/drivers/wrfda.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+CTRbin/bash
 ##################################################################################
 # Description
 ##################################################################################
@@ -52,18 +52,18 @@ fi
 ##################################################################################
 # Options below are defined in control flow xml
 #
-# N_ENS             = Max ensemble index (use 00 for control alone)
-# ANL_TIME          = Analysis time YYYYMMDDHH
-# IF_ENS_COLD_START = Skip lower / lateral BC updates if "Yes"
-# IF_LOWER          = 'Yes' if updating lower boundary conditions 
-#                     'No' if updating lateral boundary conditions
-# WRF_CTR_DOM       = Max domain index of control forecast
-# WRF_ENS_DOM       = Max domain index of ensemble perturbations
+# N_ENS         = Max ensemble index (use 00 for control alone)
+# ANL_TIME      = Analysis time YYYYMMDDHH
+# IF_ENS_UPDATE = Skip lower / lateral BC updates if "No"
+# BOUNDARY      = 'LOWER' if updating lower boundary conditions 
+#                 'LATERAL' if updating lateral boundary conditions
+# WRF_CTR_DOM   = Max domain index of control forecast
+# WRF_ENS_DOM   = Max domain index of ensemble perturbations
 #
 # Below variabs are derived by control flow variables for convenience
 #
-# date_str          = Defined by the ANL_TIME variable, to be used as path
-#                     name variable in YYYY-MM-DD_HH:MM:SS format for wrfout
+# anl_iso       = Defined by the ANL_TIME variable, to be used as path
+#                 name variable in YYYY-MM-DD_HH:MM:SS format for wrfout
 #
 ##################################################################################
 
@@ -77,24 +77,24 @@ if [ ! "${ANL_TIME}" ]; then
   exit 1
 fi
 
-# Convert ANL_TIME from 'YYYYMMDDHH' format to date_str Unix date format
+# Convert ANL_TIME from 'YYYYMMDDHH' format to anl_iso iso format
 if [ ${#ANL_TIME} -ne 10 ]; then
-  echo "ERROR: \${ANL_TIME}, '${ANL_TIME}', is not in 'yyyymmddhh' format." 
+  echo "ERROR: \${ANL_TIME}, '${ANL_TIME}', is not in 'YYYYMMDDHH' format." 
   exit 1
 else
   start_time="${START_TIME:0:8} ${START_TIME:8:2}"
   anl_date=${ANL_TIME:0:8}
   hh=${ANL_TIME:8:2}
-  date_str=`date +%Y-%m-%d_%H:%M:%S -d "${anl_date} ${hh} hours"`
+  anl_iso=`date +%Y-%m-%d_%H:%M:%S -d "${anl_date} ${hh} hours"`
 fi
 
-if [[ ${IF_ENS_COLD_START} != ${YES} && ${IF_ENS_COLD_START} != ${NO} ]]; then
-  echo "ERROR: \${IF_ENS_COLD_START} must equal 'Yes' or 'No' (case insensitive)."
+if [[ ${IF_ENS_UPDATE} != ${YES} && ${IF_ENS_UPDATE} != ${NO} ]]; then
+  echo "ERROR: \${IF_ENS_UPDATE} must equal 'Yes' or 'No' (case insensitive)."
   exit 1
 fi
 
-if [[ ${IF_LOWER} != ${YES} && ${IF_LOWER} != ${NO} ]]; then
-  echo "ERROR: \${IF_LOWER} must equal 'Yes' or 'No' (case insensitive)."
+if [[ ${BOUNDARY} != ${LOWER} && ${BOUNDARY} != ${LATERAL} ]]; then
+  echo "ERROR: \${BOUNDARY} must equal 'LOWER' or 'LATERAL' (case insensitive)."
   exit 1
 fi
 
@@ -118,6 +118,7 @@ fi
 #              vtables, geogrid data, GSI fix files, etc.
 # CYCLE_HOME = Start time named directory for cycling data containing
 #              bkg, wpsprd, realprd, wrfprd, wrfdaprd, gsiprd, enkfprd
+# ENS_ROOT   = Forecast ensemble located at ${ENS_ROOT}/ens_${ens_n}/wrfout* 
 #
 ##################################################################################
 
@@ -141,21 +142,32 @@ if [ -z ${CYCLE_HOME} ]; then
   exit 1
 fi
 
+if [ ! "${ENS_ROOT}" ]; then
+  echo "ERROR: \${ENS_ROOT} is not defined."
+  exit 1
+fi
+
+if [ ! -d "${ENS_ROOT}" ]; then
+  echo "ERROR: \${ENS_ROOT} directory ${ENS_ROOT} does not exist."
+  exit 1
+fi
+
 ##################################################################################
 # Begin pre-WRFDA setup
 ##################################################################################
 # The following paths are relative to control flow supplied root paths
 #
 # work_root      = Directory where da_update_bc.exe runs
-# gsi_dir        = Directory where GSI produces control analysis
-# enkf_dir       = Directory where EnKF produces the analysis ensemble
-# real_dir       = Directory real.exe runs and outputs IC and BC files
-# bkg_dir        = Directory with forecast data from WRF linked
+# real_dir       = Directory real.exe runs and outputs IC and BC files for cycle
+# ctr_dir        = Directory with control WRF forecast for lower boundary update 
+# ens_dir        = Directory with ensemble WRF forecast for lower boundary update 
+# gsi_dir        = Directory with GSI control analysis for lateral update
+# enkf_dir       = Directory with EnKF analysis for ensemble lateral update
 # update_bc_exe  = Path and name of the update executable
 #
 ##################################################################################
 
-if [[ ${IF_ENS_COLD_START} = ${YES} ]]; then
+if [[ ${IF_ENS_UPDATE} = ${NO} ]]; then
   # skip the boundary updates for the ensemble, perform on control alone
   n_ens=0
 else
@@ -169,11 +181,10 @@ while [ ${ens_n} -le ${n_ens} ]; do
   # define two zero padded string for GEFS 
   iimem=`printf %02d $(( 10#${ens_n} ))`
 
-  work_root=${CYCLE_HOME}/wrfdaprd/ens_${iimem}
+  work_root=${CYCLE_HOME}/wrfdaprd
+  real_dir=${CYCLE_HOME}/realprd/ens_${iimem}
   gsi_dir=${CYCLE_HOME}/gsiprd
   enkf_dir=${CYCLE_HOME}/enkfprd
-  real_dir=${CYCLE_HOME}/realprd/ens_${iimem}
-  bkg_dir=${CYCLE_HOME}/bkg/ens_${iimem}
   update_bc_exe=${WRFDA_ROOT}/var/da/da_update_bc.exe
   
   if [ ! -d ${real_dir} ]; then
@@ -186,30 +197,32 @@ while [ ${ens_n} -le ${n_ens} ]; do
     exit 1
   fi
   
-  # create working directory and cd into it
-  mkdir -p ${work_root}
-  cmd="cd ${work_root}"
-  echo ${cmd}
-  eval ${cmd}
-  
-  # Remove IC/BC in the directory if old data present
-  rm -f wrfout_*
-  rm -f wrfinput_d0*
-  rm -f wrfbdy_d01
-  
   # define domain variable to iterate on in lower boundary, fixed in lateral
   dmn=1
   
-  if [[ ${IF_LOWER} = ${YES} ]]; then 
-
+  if [[ ${BOUNDARY} = ${LOWER} ]]; then 
+    # create working directory and cd into it
+    work_root=${work_root}/lower_bdy_update/ens_${iimem}
+    mkdir -p ${work_root}
+    cmd="cd ${work_root}"
+    echo ${cmd}
+    eval ${cmd}
+    
+    # Remove IC/BC in the directory if old data present
+    rm -f wrfout_*
+    rm -f wrfinput_d0*
+  
+    # verify forecast data root
     if [ ! -d ${bkg_dir} ]; then
       echo "ERROR: \${bkg_dir} directory ${bkg_dir} does not exist."
       exit 1
     fi
     
     if [ ${ens_n} -eq 0 ]; then 
+      bkg_dir=${CYCLE_HOME}/bkg/ens_${iimem}
       max_dom=${WRF_CTR_DOM}
     else
+      bkg_dir=${ENS_ROOT}/bkg/ens_${iimem}
       max_dom=${WRF_ENS_DOM}
     fi
 
@@ -217,7 +230,7 @@ while [ ${ens_n} -le ${n_ens} ]; do
     echo "Copying background and input files."
     while [ ${dmn} -le ${max_dom} ]; do
       # update the lower BC for the output file to pass to GSI
-      wrfout=wrfout_d0${dmn}_${date_str}
+      wrfout=wrfout_d0${dmn}_${anl_iso}
 
       # wrfinput is always drawn from real step
       wrfinput=wrfinput_d0${dmn}
@@ -284,8 +297,9 @@ while [ ${ens_n} -le ${n_ens} ]; do
       echo "WRFDA_ROOT = ${WRFDA_ROOT}"
       echo "EXP_CONFIG = ${EXP_CONFIG}"
       echo "CYCLE_HOME = ${CYCLE_HOME}"
+      echo "ENS_ROOT   = ${ENS_ROOT}"
       echo
-      echo "IF_LOWER   = ${IF_LOWER}"
+      echo "BOUNDARY   = ${BOUNDARY}"
       echo "DOMAIN     = ${dmn}"
       echo "ENS_N      = ${ens_n}"
       echo
@@ -302,219 +316,127 @@ while [ ${ens_n} -le ${n_ens} ]; do
         echo "ERROR: ${update_bc_exe} exited with status ${error}."
         exit ${error}
       fi
-  
-      # save the files where they can be accessed for GSI analysis
-      lower_bdy_data=${work_root}/lower_bdy_update
-      mkdir -p ${lower_bdy_data}
-      mv ${wrfout} ${lower_bdy_data}/${wrfout}
-      mv ${wrfinput} ${lower_bdy_data}/${wrfinput}
-      mv parame.in ${lower_bdy_data}/parame.in_d0${dmn} 
   
       # move forward through domains
       (( dmn += 1 ))
     done
   
   else
+    # create working directory and cd into it
+    work_root=${work_root}/lateral_bdy_update/ens_${iimem}
+    mkdir -p ${work_root}
+    cmd="cd ${work_root}"
+    echo ${cmd}
+    eval ${cmd}
+    
+    # Remove IC/BC in the directory if old data present
+    rm -f wrfout_*
+    rm -f wrfinput_d0*
+    rm -f wrfbdy_d01
+
     if [ ${ens_n} -eq 0 ]; then
       if [ ! -d ${gsi_dir} ]; then
         echo "ERROR: \${gsi_dir} directory ${gsi_dir} does not exist."
         exit 1
-      fi
-
-      wrfanl=${gsi_dir}/d01/wrfanl_ens_00.${ANL_TIME}
-      wrfbdy=${real_dir}/wrfbdy_d01
-      wrfvar_outname=wrfvar_output
-      wrfbdy_name=wrfbdy_d01
-  
-      if [ ! -r "${wrfanl}" ]; then
-        echo "ERROR: Input file '${wrfanl}' is missing."
-        exit 1
       else
-        cmd="cp ${wrfanl} ${wrfvar_outname}"
-	echo ${cmd}
-	eval ${cmd}
+        wrfanl=${gsi_dir}/d01/wrfanl_ens_${ens_n}_${anl_iso}
       fi
-  
-      if [ ! -r "${wrfbdy}" ]; then
-        echo "ERROR: Input file '${wrfbdy}' is missing."
-        exit 1
-      else
-        cmd="cp ${wrfbdy} ${wrfbdy_name}"
-	echo ${cmd}
-	eval ${cmd}
-      fi
-  
-      ##################################################################################
-      #  Build da_update_bc namelist
-      ##################################################################################
-      # Copy the namelist from the static dir -- THIS WILL BE MODIFIED DO NOT LINK TO IT
-      cmd="cp ${EXP_CONFIG}/namelists/parame.in ./"
-      echo ${cmd}
-      eval ${cmd}
-  
-      # Update the namelist for the domain id 
-      cat parame.in \
-	 | sed "s/\(${DA}_${FILE}\)${EQUAL}.*/\1 = '\.\/${wrfvar_outname}'/" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      cat parame.in \
-	 | sed "s/\(${DOMAIN}_${ID}\)${EQUAL}[[:digit:]]\{1,\}/\1 = ${dmn}/" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      # Update the namelist for lower boundary update 
-      cat parame.in \
-	 | sed "s/\(${WRF}_${BDY}_${FILE}\)${EQUAL}.*/\1 = '\.\/${wrfbdy_name}'/" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      cat parame.in \
-	 | sed "s/\(${UPDATE}_${LOW}_${BDY}\)${EQUAL}.*/\1 = \.false\./" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      cat parame.in \
-	 | sed "s/\(${UPDATE}_${LATERAL}_${BDY}\)${EQUAL}.*/\1 = \.true\./" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      ##################################################################################
-      # Run update_bc_exe
-      ##################################################################################
-      # Print run parameters
-      echo
-      echo "WRFDA_ROOT = ${WRFDA_ROOT}"
-      echo "EXP_CONFIG = ${EXP_CONFIG}"
-      echo "EXP_CONFIG = ${EXP_CONFIG}"
-      echo "CYCLE_HOME = ${CYCLE_HOME}"
-      echo
-      echo "IF_LOWER   = ${IF_LOWER}"
-      echo "DOMAIN     = ${dmn}"
-      echo "ENS_N      = ${ens_n}"
-      echo
-      now=`date +%Y%m%d%H%M%S`
-      echo "da_update_bc.exe started at ${now}."
-      ${update_bc_exe}
-  
-      ##################################################################################
-      # Run time error check
-      ##################################################################################
-      error=$?
-      
-      if [ ${error} -ne 0 ]; then
-        echo "ERROR: ${update_bc_exe} exited with status ${error}."
-        exit ${error}
-      fi
-  
-      # save the files where they can be accessed for new WRF forecast
-      lateral_bdy_data=${work_root}/lateral_bdy_update
-      mkdir -p ${lateral_bdy_data}
-      mv wrfvar_output ${lateral_bdy_data}/
-      mv wrfbdy_d01 ${lateral_bdy_data}/
-      mv parame.in ${lateral_bdy_data}/parame.in_d0${dmn} 
-      mv fort.* ${lateral_bdy_data}/
-
     else
       if [ ! -d ${enkf_dir} ]; then
         echo "ERROR: \${enkf_dir directory} ${enkf_dir} does not exist."
         exit 1
+      else
+        # NOTE: ENKF SCRIPT NEED TO UPDATE OUTPUT NAMING CONVENTIONS
+        wrfanl=${enkf_dir}/d01/wrfanl_ens_${ens_n}_${anl_iso}
       fi
+    fi
 
-      wrfanl=${enkf_dir}/d01/analysis.mem${iimem}
-      wrfbdy=${real_dir}/wrfbdy_d01
-      wrfvar_outname=wrfvar_output
-      wrfbdy_name=wrfbdy_d01
+    wrfbdy=${real_dir}/wrfbdy_d01
+    wrfvar_outname=wrfanl_ens_${ens_n}_${anl_iso}
+    wrfbdy_name=wrfbdy_d01
   
-      if [ ! -r "${wrfanl}" ]; then
-        echo "ERROR: Input file '${wrfanl}' is missing."
-        exit 1
-      else
-        cmd="cp ${wrfanl} ${wrfvar_outname}"
-	echo ${cmd}
-	eval ${cmd}
-      fi
-  
-      if [ ! -r "${wrfbdy}" ]; then
-        echo "ERROR: Input file '${wrfbdy}' is missing"
-        exit 1
-      else
-        cmd="cp ${wrfbdy} ${wrfbdy_name}"
-	echo ${cmd}
-	eval ${cmd}
-      fi
-  
-      ##################################################################################
-      #  Build da_update_bc namelist
-      ##################################################################################
-      # Copy the namelist from the static dir -- THIS WILL BE MODIFIED DO NOT LINK TO IT
-      cmd="cp ${EXP_CONFIG}/namelists/parame.in ./"
+    if [ ! -r "${wrfanl}" ]; then
+      echo "ERROR: Input file '${wrfanl}' is missing."
+      exit 1
+    else
+      cmd="cp ${wrfanl} ${wrfvar_outname}"
       echo ${cmd}
       eval ${cmd}
-  
-      # Update the namelist for the domain id 
-      cat parame.in \
-	 | sed "s/\(${DA}_${FILE}\)${EQUAL}.*/\1 = '\.\/${wrfvar_outname}'/" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      cat parame.in \
-	 | sed "s/\(${DOMAIN}_${ID}\)${EQUAL}[[:digit:]]\{1,\}/\1 = ${dmn}/" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      # Update the namelist for lower boundary update 
-      cat parame.in \
-	 | sed "s/\(${WRF}_${BDY}_${FILE}\)${EQUAL}.*/\1 = '\.\/${wrfbdy_name}'/" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      cat parame.in \
-	 | sed "s/\(${UPDATE}_${LOW}_${BDY}\)${EQUAL}.*/\1 = \.false\./" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      cat parame.in \
-	 | sed "s/\(${UPDATE}_${LATERAL}_${BDY}\)${EQUAL}.*/\1 = \.true\./" \
-         > parame.in.new
-      mv parame.in.new parame.in
-  
-      ##################################################################################
-      # Run update_bc_exe
-      ##################################################################################
-      # Print run parameters
-      echo
-      echo "WRFDA_ROOT = ${WRFDA_ROOT}"
-      echo "EXP_CONFIG = ${EXP_CONFIG}"
-      echo "EXP_CONFIG = ${EXP_CONFIG}"
-      echo "CYCLE_HOME = ${CYCLE_HOME}"
-      echo
-      echo "IF_LOWER   = ${IF_LOWER}"
-      echo "DOMAIN     = ${dmn}"
-      echo "ENS_N      = ${ens_n}"
-      echo
-      now=`date +%Y%m%d%H%M%S`
-      echo "da_update_bc.exe started at ${now}."
-      ${update_bc_exe}
-  
-      ##################################################################################
-      # Run time error check
-      ##################################################################################
-      error=$?
-      
-      if [ ${error} -ne 0 ]; then
-        echo "ERROR: ${update_bc_exe} exited with status ${error}."
-        exit ${error}
-      fi
-  
-      # save the files where they can be accessed for new WRF forecast
-      lateral_bdy_data=${work_root}/lateral_bdy_update
-      mkdir -p ${lateral_bdy_data}
-      mv wrfvar_output ${lateral_bdy_data}/
-      mv wrfbdy_d01 ${lateral_bdy_data}/
-      mv parame.in ${lateral_bdy_data}/parame.in_d0${dmn} 
-      mv fort.* ${lateral_bdy_data}/
     fi
+  
+    if [ ! -r "${wrfbdy}" ]; then
+      echo "ERROR: Input file '${wrfbdy}' is missing."
+      exit 1
+    else
+      cmd="cp ${wrfbdy} ${wrfbdy_name}"
+      echo ${cmd}
+      eval ${cmd}
+    fi
+  
+    ##################################################################################
+    #  Build da_update_bc namelist
+    ##################################################################################
+    # Copy the namelist from the static dir -- THIS WILL BE MODIFIED DO NOT LINK TO IT
+    cmd="cp ${EXP_CONFIG}/namelists/parame.in ./"
+    echo ${cmd}
+    eval ${cmd}
+  
+    # Update the namelist for the domain id 
+    cat parame.in \
+       | sed "s/\(${DA}_${FILE}\)${EQUAL}.*/\1 = '\.\/${wrfvar_outname}'/" \
+       > parame.in.new
+    mv parame.in.new parame.in
+  
+    cat parame.in \
+       | sed "s/\(${DOMAIN}_${ID}\)${EQUAL}[[:digit:]]\{1,\}/\1 = ${dmn}/" \
+       > parame.in.new
+    mv parame.in.new parame.in
+  
+    # Update the namelist for lower boundary update 
+    cat parame.in \
+       | sed "s/\(${WRF}_${BDY}_${FILE}\)${EQUAL}.*/\1 = '\.\/${wrfbdy_name}'/" \
+       > parame.in.new
+    mv parame.in.new parame.in
+  
+    cat parame.in \
+       | sed "s/\(${UPDATE}_${LOW}_${BDY}\)${EQUAL}.*/\1 = \.false\./" \
+       > parame.in.new
+    mv parame.in.new parame.in
+  
+    cat parame.in \
+       | sed "s/\(${UPDATE}_${LATERAL}_${BDY}\)${EQUAL}.*/\1 = \.true\./" \
+       > parame.in.new
+    mv parame.in.new parame.in
+  
+    ##################################################################################
+    # Run update_bc_exe
+    ##################################################################################
+    # Print run parameters
+    echo
+    echo "WRFDA_ROOT = ${WRFDA_ROOT}"
+    echo "EXP_CONFIG = ${EXP_CONFIG}"
+    echo "EXP_CONFIG = ${EXP_CONFIG}"
+    echo "CYCLE_HOME = ${CYCLE_HOME}"
+    echo "ENS_ROOT   = ${ENS_ROOT}"
+    echo
+    echo "BOUNDARY   = ${BOUNDARY}"
+    echo "DOMAIN     = ${dmn}"
+    echo "ENS_N      = ${ens_n}"
+    echo
+    now=`date +%Y%m%d%H%M%S`
+    echo "da_update_bc.exe started at ${now}."
+    ${update_bc_exe}
+  
+    ##################################################################################
+    # Run time error check
+    ##################################################################################
+    error=$?
+    
+    if [ ${error} -ne 0 ]; then
+      echo "ERROR: ${update_bc_exe} exited with status ${error}."
+      exit ${error}
+    fi
+    
   fi
 
   (( ens_n += 1 ))
